requirements.txt
absl-py==2.1.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==1.3.0
altair==4.2.2
annotated-types==0.7.0
annotator==0.14.2
antlr4-python3-runtime==4.8
anyio==4.5.2
async-timeout==5.0.1
attrs==25.1.0
backports.zoneinfo==0.2.1
basicsr==1.4.2
blinker==1.8.2
braceexpand==0.1.7
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1725551672989/work/certifi
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
config==0.5.1
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.3.0
elasticsearch==1.9.0
entrypoints==0.4
eval-type-backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.2.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
gradio==3.16.2
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna @ file:///croot/idna_1714398848350/work
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
invisible-watermark==0.2.0
iso8601==2.1.0
jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kornia==0.6.0
lazy-loader==0.4
lightning-utilities==0.11.9
linkify-it-py==2.0.3
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.29.0
networkx==3.1
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch @ git+https://github.com/mlfoundations/open_clip.git@a09d519f91358187d73debacfc2db5f154291d40
opencv-contrib-python==4.3.0.36
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.0.3
pillow @ file:///croot/pillow_1721059439630/work
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prettytable==3.6.0
propcache==0.2.0
protobuf==3.20.3
psutil==7.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
pygments==2.19.1
PyJWT==2.9.0
Pympler==1.1
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytorch-lightning==1.5.0
pytz==2025.1
PyWavelets==1.4.1
PyYAML==6.0.2
qudida==0.0.4
referencing==0.35.1
regex==2024.11.6
requests @ file:///croot/requests_1721410876868/work
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
share==1.0.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
streamlit-drawable-canvas==0.8.0
sympy==1.13.3
tb-nightly==2.14.0a20230808
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.12
tokenizers==0.20.3
toml==0.10.2
toolz==1.0.0
torch==2.4.1
torchaudio==2.4.1
torchmetrics==1.5.2
torchvision==0.20.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing-extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uc-micro-py==1.0.3
urllib3==1.26.20
uvicorn==0.33.0
validators==0.34.0
wandb==0.19.7
watchdog==4.0.2
wcwidth==0.2.13
webdataset==0.2.5
websockets==13.1
werkzeug==3.0.6
xformers==0.0.28.post1
yapf==0.32.0
yarl==1.15.2
zipp==3.20.2
